requirements.txt
appnope==0.1.3
asttokens==2.2.1
autograd==1.5
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
escnn==1.0.7
executing==1.2.0
filelock==3.11.0
fonttools==4.39.3
future==0.18.3
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lie-learn==0.0.1.post1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py3nj==0.1.2
Pygments==2.15.0
pymanopt==2.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==25.0.2
requests==2.28.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
torch==2.0.0
torchinfo==1.7.2
torchsummary==1.5.1
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.15.0
